+++ packages/ciphernode/p2p/src/lib.rs
@@ -1,8 +1,8 @@
 #![crate_name = "p2p"]
 #![crate_type = "lib"]
 
-mod libp2p_router;
+mod network_peer;
 mod p2p;
 
-pub use libp2p_router::*;
+pub use network_peer::*;
 pub use p2p::*;
